VPP Python API invocation mechanism has changed in 18.04
VPP version 18.04 onwards, the VPP APIs are attributes of the "api"
object within the VPPInterface object whereas before 18.04, VPP APIs
are attributes of the VPPInterface object itself. As a result, none of the
VPP API calls work in the ML2 agent when run against VPP version 18.04.

This change ensures that we work with VPP version 18.04 and onwards while
still being backwards compatible.

Change-Id: Iac4d41a4295cfbd6985a8f0f2595e73e141a116c
Closes-Bug: #1756051