L3 enhancement - external network connectivity and FIP
Connectivity to the external network is now handled by the networking-vpp
L3 agent automatically, i.e., when external gateway is set for a router,
the L3 agent looks for an external linux bridge, br-ex, on the
controller/network node(s) and connects VPP to it using a veth pair.
Subsequently, tenant instances can access the external network (SNAT) and
can be accessed from outside through FIP seamlessly. The only manual step
required would be the setting up of the external linux bridge on the
controller/network node(s).

NOTE: A different bridge name can be specified in the configuration file
with the new “l3_external_bridge” option.

Change-Id: I4e237cc7bc4f4a579255a237d9719fdacf5d8531