Fix agent crash during GPE port binding.
The vpp-agent crashes during GPE port binding because it tries to add an
incorrect MAC address to the GPE mapping data. We are using a dynamically
generated MAC address for binding ports. To resolve this issue, we must fetch
the correct mac address of the port and use it for GPE.

Closes-Bug: #1760726

Change-Id: I512dcf8bfbe7b4a5df0f9f773eecb7f06177c244