Update the GPE etcd key format to handle vpp-agent restarts correctly
Correctly handle mac-address changes during vpp-agent restarts
by removing the mac-address of a port from the etcd key.
The port mac-address could be computed dynamically by the vpp-agent
and could change after an agent restart. If the mac-address is
present in the etcd key, it results in duplicate GPE keys and GPE
stops working. This patch resolves this issue by removing the
mac-address from the key and storing it in the value.

Closes-Bug: #1761023

Change-Id: I3f1f4168f0c77080edc076bbb660021b81fb62dc