Remove anti-pattern from etcd signing code
The 'SignedEtcdJSONWriter' is, depending on config, a 'signed etcd json
writer' or 'a proxy to the unsigned etcd json writer'.  The two classes
should be instances of the same base type and the correct one
instantiated based on circumstances.

This change converts users of the JSON writers to use a function that
instantiates the right JSON writer based on circumstances.

Change-Id: If01c68d93d06c62bea8e55568d5049a3aa8c2c97