Fix issue in device_monitor eventlet
pyudev <= v0.15 does not support the poll() method and this causes the
device_monitor eventlet to die with the following exception:

AttributeError: 'Monitor' object has no attribute 'poll'

As a result, deployments which have pyudev <= v0.15, will see intermittent
issues with DHCP. Instances will fail to obtain IP addr from DHCP when there
is a slight delay in the kernel tap creation by the neutron DHCP agent.

The fix ensures that the agent is able to handle pyudev <= v0.15 as well as
pyudev > 0.15.

Change-Id: I1f6270eb843c945eb1ad5a90a8dc5226a66c9d19
Closes-Bug: #1763814