Return interface info with existing bindings
The current networking-vpp has two sources for binding interfaces,
either from regular NEUTRON binding requests, or from file monitor in
the case of live migration. The interface binding function is
designed to handle the case with a lock, but needs to return the
existing binding information if the binding is already done.

Change-Id: I05d5b4dc1ea9ab3dc020c3ca3a6a07b240450929