Convert tap_connect to tapv2 APIs
The older TAP drivers and their APIs have been deprecated in favour of
the v2 APIs.  This converts vpp.py to call those while not changing the
code of the agent.

The devstack plugin, where it uses a TAP interface, has also been
adjusted to use the new APIs and the new (and slightly different) form
of TAP interface name in VPP.

Change-Id: I7c713eab3c864f871de98b2cd975bf182f4def35