Change the way security group options are registered
We used to register them using the Neutron-internal list of
information, but this actually includes options we don't implement
and it is, in any case, not appropriate for us to assume that the
list appropriate to Neutron's internal agents is appropriate to
our mechanism driver.  Use the local list but adapt in case someone
or something has registered the config item before.

Change-Id: Ia0ec62823705bd5027bf1bf46c546f4dab92bf19