Polling Linux for vhost socket status instead of from VPP
We need to detect when new vhost-user ports have connected to their
VMs.  We did this by listing all vhostuser ports and checking them, 
since there's no notification API, but it turns out that VPP locking
can sometimes prevent packet forwarding during the API call.  (This
is obvious with a combination of the 18.07 PAPI library, 32+ ports
and high throughput, but can occur in other circumstances.)

To avoid this problem, change the vhost-user polling to check the
connectivity of sockets via the kernel.  This is just as effective
and does not interact with VPP at all.

Change-Id: Ie10df10df995ed5232b15e7d620dd09508aac695