etcd refresh issue
Older etcd client implementations, which do not support the refresh
method, result in:

AttributeError: 'Client' object has no attribute 'refresh'

This leads to the eventual deletion of the node's aliveness key in etcd
and which in turn results in port creation/deletion failure.

The fix is to invoke the etcd client's write method with the refresh
flag set (which is what the refresh method does too). This ensures
that we are not hampered by older versions of etcd client.

Change-Id: I5056469d6368e15b250a361ee773e5249a271a10
Closes-Bug: #1783548