sw_interface_set_mtu has changed in VPP 18.07
In VPP 18.07, the API's MTU field is an array which allows for setting
MTUs for generic L3, IPv4, IPv6 and MPLS:

    u32 mtu[4]; /* 0 - L3, 1 - IP4, 2 - IP6, 3 - MPLS */

Details in the following link:
    https://docs.fd.io/vpp/18.07/md_src_vnet_MTU.html

As a result, ML2 agent fails to set the MTU for the router's
interfaces. The fix is to set the generic L3 MTU and leave the
protocol specific MTU fields alone.

Change-Id: I215416037d727e887aa44453437f1b90129cb172
Closes-Bug: #1784642