timing issues in trunk port feature
During restart/resync, the reconsider_trunk_subports() function is
invoked only once. As a result, child ports may never be bound if
there's a lag in the parent port’s bind.

A second related issue is when the parent port is bound but not 'up'
yet then the child subport stays in the 'down' state. A side-effect of
this issue is that trunk port artifacts in VPP are not cleaned up upon
unbind of the parent port.

Change-Id: I03839f36b7063b92d1420ae7a91e5841f3623018
Closes-Bug: #1788848