Replace device monitor
The pyudev device monitor is quite nasty - pyudev is not eventlet
friendly and needs a lot of workarounds to operate, and we have
reports of the callback eventlet locking up.

Replace the pyudev dependency with some homebrew netlink code.
(There are a few Python libraries for netlink out there, including
pyroute2, but none that claim to be better than beta quality.  The
use of netlink is fairly straightforward, so the code here is not
huge.)

TODO: when doing the initial sync there is some disagreement about
the size of data provided by netlink as a single message and the
size of the data structure in its header, causing warnings, and this
needs to be investigated.

Change-Id: Ifc058c9cf3f7cf2aae718893735467ada1f165f3
Co-Authored-By: Hareesh Puthalath <hputhala@cisco.com>