Remove unnecessary SQLAlchemy flush() calls
session.flush() converts the current in-memory data changes to DB
changes.  However, it misleadingly stackdumps if there's a problem with
*any* DB statement so far, which leads to misleading net-vpp backtraces
when the actual table involved may have nothing to do with net-vpp.
It's also intended to flush out huge transactions in pieces, which we
don't have; so remove the call.

Change-Id: Ic9f81890fb7858f6922575d4aa1792d4e4f4cac8