Initialize privsep in vpp-agent
We were not initializing privsep in the vpp-agent so we weren't
using any root helper to call it. In devstack this is not a problem
since 'stack' user can execute anything but in TripleO or OPNFV Fuel
deployments it won't work since user 'neutron' is only allowed to
execute commands through rootwrap/rootwrap-daemon.

Closes-bug: #1809887

Change-Id: I48cf586fa3a4118c9ab2f7b90ca7848f3f237e92
Signed-off-by: Alexandru Avadanii <Alexandru.Avadanii@enea.com>