Support for multiple router gateways on the floating network.
Since the VPP does not support overlapping IP addresses on a
subnet, the solution adds an independent route to the VPP's
local interface each time an additional router is added on the
floating network. The first router's gateway IP address is set
on the BVI loopback and this address becomes the primary gateway
IP address. Each subsequent router's gateway IP address is added
as a local route. When the primary gateway is deleted, if a
valid local IP address exists, it is migrated to the BVI
loopback interface. Upon deleting an external gateway, the BVI
loopback is only deleted if no valid local IPs exist on the
floating network.

Change-Id: I945056fad31e599833e706ee627fe444e32ed606
Closes-Bug: #1810536